src/users/models.py
import uuid
from django.db import models
from django.dispatch import receiver
from django.contrib.auth.models import AbstractUser
from rest_framework_simplejwt.tokens import RefreshToken
from easy_thumbnails.fields import ThumbnailerImageField
from django.urls import reverse
from django_rest_passwordreset.signals import reset_password_token_created
from easy_thumbnails.signals import saved_file
from easy_thumbnails.signal_handlers import generate_aliases_global
from phonenumber_field.modelfields import PhoneNumberField

from src.common.helpers import build_absolute_uri
from src.notifications.services import notify, ACTIVITY_USER_RESETS_PASS


@receiver(reset_password_token_created)
def password_reset_token_created(sender, instance, reset_password_token, *args, **kwargs):
    """
    Handles password reset tokens
    When a token is created, an e-mail needs to be sent to the user
    """
    reset_password_path = reverse('password_reset:reset-password-confirm')
    context = {
        'username': reset_password_token.user.username,
        'email': reset_password_token.user.email,
        'reset_password_url': build_absolute_uri(f'{reset_password_path}?token={reset_password_token.key}'),
    }

    notify(ACTIVITY_USER_RESETS_PASS, context=context, email_to=[reset_password_token.user.email])


class User(AbstractUser):
    id = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
    other_name = models.CharField(max_length=50, blank=True, null=True)
    gender = models.CharField(default='prefer not to say', max_length=20, blank=True, null=True)
    date_of_birth = models.CharField(max_length=10, blank=True, null=True)
    phone_number = PhoneNumberField(blank=True, null=True)
    house_address = models.CharField(max_length=500, blank=True, null=True)
    city = models.CharField(max_length=20, blank=True, null=True) 
    state = models.CharField(max_length=30, blank=True, null=True)
    email = models.CharField(max_length=100, blank=True, null=True)
    bvn = models.PositiveBigIntegerField(blank=True, null=True)
    is_bvnverified = models.BooleanField(default=False, null=True)#Will remain False until bvn is verified
    is_emailverified = models.BooleanField(default=False, null=True)#Will remain False until email is verified
    profile_picture = ThumbnailerImageField('ProfilePicture', upload_to='profile_pictures/', blank=True, null=True)

    def get_tokens(self):
        refresh = RefreshToken.for_user(self)

        return {
            'refresh': str(refresh),
            'access': str(refresh.access_token),
        }

    def __str__(self):
        return self.username


saved_file.connect(generate_aliases_global)
